+++ src/main/java/br/com/pi/projeto_RD/model/dto/ItensDfDTO.java
@@ -3,23 +3,18 @@ package br.com.pi.projeto_RD.model.dto;
 import br.com.pi.projeto_RD.model.entity.ProdutoFilialEstoqueEntity;
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Data
 public class ItensDfDTO {
 
-    private Long nrItemDocumento;
-
-    private Integer cdProduto;
+    private BigInteger nrItemDocumento;
+    private BigInteger cdProduto;
     private String nmProduto;
-
     private Integer qtItem;
-
     private Double vlItem;
     private Double pcIcms;
     private Double vlIcms;
-
-//    private List<ProdutoFilialEstoqueEntity> produtoEstoque;
-
 
 }
