+++ src/main/java/br/com/pi/projeto_RD/repository/DocumentoFiscalRepository.java
@@ -10,10 +10,11 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
-public interface DocumentoFiscalRepository extends JpaRepository<DocumentoFiscalEntity, Long> {
+public interface DocumentoFiscalRepository extends JpaRepository<DocumentoFiscalEntity, BigInteger> {
 
     @EntityGraph(value = "documentosfindall", type = EntityGraph.EntityGraphType.LOAD)
     List<DocumentoFiscalEntity> findAll();
